Course Assignments/CP8319 - Reinforcement Learning Assignments/CP8319 Assignment 3/src/requirements.txt
gym
atari-py
matplotlib
numpy
scipy
pytorch
torchvision
torchaudio
tensorboard